Feat: Add JPA Annotations to Task Model
